Rename Command to ICommand
